Rename folders to include 08_icons
