+++ src/04_templates/main-layout-template.js
@@ -3,7 +3,7 @@
  */
 
 import React from "react"
-import "../09_styles/styles.scss"
+import "../10_styles/styles.scss"
 import Header from "../03_organisms/header"
 import Footer from "../03_organisms/footer"
 import { SkipNavLink, SkipNavContent } from "@reach/skip-nav"
