+++ src/04_templates/post-template.js
@@ -6,7 +6,7 @@ import React from "react"
 import { graphql } from "gatsby"
 import { Link } from "gatsby"
 import MainLayout from "../04_templates/main-layout-template"
-import SEO from "../08_utilities/seo"
+import SEO from "../09_utilities/seo"
 
 export default function Template({
     data, // This prop will be injected by the GraphQL query below.
