+++ src/10_styles/04_template-styles/_main-layout-template.scss
